#1826 - fix(design-system): move buttonVariants out of the component (1.0.1)
buttonVariants is a class contract, not a component, but it lived in
button.svelte's <script module>. That meant non-component code wanting the
class strings had to import a .svelte file — and importing it via the /ui
barrel pulled in every primitive, Tooltip's bits-ui dependency included.

That's a lot of module graph for a string of Tailwind classes. It costs every
consumer bundle size, and it broke a real test: saiku-cloud's tour.test.ts
dynamically imports src/lib/onboarding/tour.ts, which needs buttonVariants to
style driver.js buttons. When that import went from button.svelte to the
barrel, the transform cost tipped the test past its 5s timeout on CI's slower
runners. It passes locally on a fast machine either way, which is exactly the
kind of regression that only shows up in CI.

Extracted to ui/button-variants.ts — a plain module whose only import is
tailwind-variants. button.svelte re-exports it and the /ui barrel now sources
it from there, so `import { buttonVariants } from '.../ui'` is unchanged for
existing callers. New ./ui/button-variants subpath export is the cheap path
for anything that only wants the classes.

Additive only, no breaking change.

Verified: saiku-ui 0 svelte-check errors, 2119 tests green; package
type-checks standalone with 0 errors.
