+++ saiku-ui/design-system/src/lib/ui/index.ts
@@ -1,5 +1,12 @@
-export { default as Button, buttonVariants } from './button.svelte';
-export type { ButtonVariant, ButtonSize } from './button.svelte';
+export { default as Button } from './button.svelte';
+// Variants come straight from the .ts module, not through button.svelte — this
+// barrel is a convenience, and routing the class contract through a component
+// would make `import { buttonVariants }` drag the component graph in. Anything
+// that ONLY wants the classes should import '.../ui/button-variants' directly
+// and skip this barrel (which pulls in every primitive, Tooltip's bits-ui
+// included).
+export { buttonVariants } from './button-variants';
+export type { ButtonVariant, ButtonSize } from './button-variants';
 export { default as Card } from './card.svelte';
 export { default as CardHeader } from './card-header.svelte';
 export { default as CardTitle } from './card-title.svelte';
